+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - [#55](https://github.com/wandersoncferreira/code-review/pull/55): fix `(wrong-type-argument (or eieio-object class) nil obj)`. Special thanks to @ktfleming
 - [#56](https://github.com/wandersoncferreira/code-review/pull/56): remove `code-review-submit` interactive call from transient options.
 - [#57](https://github.com/wandersoncferreira/code-review/pull/57): fix error when adding a comment after a reply obj
+- [#61](https://github.com/wandersoncferreira/code-review/pull/61): allow approve PR without feedback comment